requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Pillow==10.1.0
PyMuPDF==1.23.8
python-docx==1.1.0
RPi.GPIO==0.7.1
requests==2.31.0
